stdb/requirement.txt
asgiref==3.7.2
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
distlib==0.3.6
distro==1.8.0
Django==4.2.2
django-environ==0.10.0
django-filter==23.2
django-perms==2.1.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==6.1.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.6
environs==9.5.0
filelock==3.12.2
generics==6.0.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
marshmallow==3.19.0
nose==1.3.7
packaging==23.1
paramiko==3.2.0
pathlib==1.0.1
phonenumbers==8.13.14
Pillow==9.5.0
pipenv==2023.6.18
platformdirs==3.7.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
PyYAML==6.0
requests==2.31.0
rest-framework-simplejwt==0.0.2
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
texttable==1.6.7
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
virtualenv==20.23.1
virtualenv-clone==0.5.7
websocket-client==1.6.0
xlwt==1.3.0
